Added tests of rest client

+++ src/main/java/io/dbeauregard/pivotalspring_client/House.java
@@ -3,4 +3,4 @@ package io.dbeauregard.pivotalspring_client;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record House(Long id, String address, Long price) {}
+public record House(Long id, String address, Integer price) {}
